tech stack component
