+++ src/pages/me.js
@@ -1,8 +1,10 @@
+/* eslint-disable prettier/prettier */
 import React from "react";
 import PropTypes from "prop-types";
 import { css } from "@emotion/core";
-import { graphql } from "gatsby";
+import { graphql, StaticQuery } from "gatsby";
 import Img from "gatsby-image";
+import Techstack from "../components/Techstack";
 import Layout from "../components/layout";
 
 const styles = {
@@ -32,38 +34,83 @@ export const query = graphql`
         }
       }
     }
+    site {
+      siteMetadata {
+        techstack
+      }
+    }
   }
 `;
 
 function Me(props) {
-  const { location, data } = props;
+  const { location } = props;
   return (
-    <Layout location={location} maxWidth="650px">
-      <article>
-        <div />
-        <h1 css={styles.title}>Hello, My name is Hans</h1>
-        <Img css={styles.img} fluid={data.fileName.childImageSharp.fluid} />
-        <div css={styles.content}>
-          <p>
-            Hello world, my name is Hans Garcia, I am a computer science
-            engineer from Panama, I've been programming for +6 years as
-            fullstack developer with frontend tendencies, the last two years I
-            have been using <strong>react.js</strong>,
-            <strong>react-native</strong>,{" "}
-            <strong>firebase/firestore/cloud functions</strong>,{" "}
-            <strong>postgresql</strong>, <strong>node.js and aws lambda functions</strong>
-          </p>
-          <p>
-            In the pass I worked with <strong>.net core</strong>, <strong>knockout.js</strong> and <strong>Java</strong>. I think
-            the present/future is serverless side
-          </p>
-          <p>
-            I love pizza (with pineapple plis), coffee and I prefer tabs over
-            space
-          </p>
-        </div>
-      </article>
-    </Layout>
+    <StaticQuery
+      query={query}
+      render={({ site, fileName }) => {
+        console.log(site, fileName)
+        const { techstack } = site.siteMetadata;
+        return (
+          <Layout location={location} maxWidth="650px">
+            <article>
+              <h1 css={styles.title}>Hello, My name is Hans</h1>
+              <div css={css`
+          displa: flex;
+        `}
+              >
+                <div>
+                  <Img css={styles.img} fluid={fileName.childImageSharp.fluid} />
+                  <Techstack stack={techstack} />
+                </div>
+                <div css={styles.content}>
+                  <p>
+              Hello world, my name is Hans Garcia, I am a computer science
+              engineer from Panama, I've been programming for +6 years as
+              fullstack developer with frontend tendencies, the last two years I
+              have been using
+                    {' '}
+                    <strong>react.js</strong>
+,
+                    {' '}
+                    <strong>react-native</strong>
+,
+                    {' '}
+                    {" "}
+                    <strong>firebase/firestore/cloud functions</strong>
+,
+                    {" "}
+                    <strong>postgresql</strong>
+,
+                    {" "}
+                    <strong>node.js and aws lambda functions</strong>
+                  </p>
+                  <p>
+              In the pass I worked with
+                    {' '}
+                    <strong>.net core</strong>
+,
+                    {" "}
+                    <strong>knockout.js</strong>
+                    {' '}
+and
+                    {' '}
+                    <strong>Java</strong>
+. I think
+              the present/future is serverless side
+                  </p>
+                  <p>
+              I love pizza (with pineapple plis), coffee and I prefer tabs over
+              space
+                  </p>
+                </div>
+              </div>
+            </article>
+          </Layout>
+)
+      }}
+    />
+
+
   );
 }
